refactor(animations): moved animations into own class
